Generalize SSE to different events
