+++ backend/config/config.yaml
@@ -7,5 +7,5 @@ instruments:
       ADM, ANET, AJG, AIZ, T, ATO, ADSK, ADP, 
     ]
 logging:
-  level: "WARNING"
+  level: "DEBUG"
   format: "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
