+++ backend/config/logging.py
@@ -21,19 +21,6 @@ def setup_logging() -> None:
         format=logging_config.get("format"),
     )
 
-    # Set yfinance logger to only show WARNING and above
-    logging.getLogger("modules.prices.services").setLevel(logging.DEBUG)
-    # logging.getLogger("modules.sse.sse_consumer_impl").setLevel(logging.DEBUG)
-    # logging.getLogger("modules.prices.live_prices_service").setLevel(logging.DEBUG)
-    logging.getLogger("django.request").setLevel(logging.WARNING)
-    logging.getLogger("yfinance").setLevel(logging.DEBUG)
-    logging.getLogger("websockets.client").setLevel(logging.WARNING)
-    logging.getLogger("httpcore.http11").setLevel(logging.WARNING)
-    logging.getLogger("httpx").setLevel(logging.WARNING)
-    logging.getLogger("daphne.http_protocol").setLevel(logging.WARNING)
-    logging.getLogger("daphne.server").setLevel(logging.WARNING)
-    logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
-
 
 def get_logger(name: str) -> logging.Logger:
     """Get a logger instance with the given name."""
