+++ backend/modules/prices/containers.py
@@ -1,10 +1,11 @@
 from dependency_injector import containers, providers
 
 from modules.prices.repositories import YfinanceRepository
-from modules.prices.services import LivePricesService, PricesService
+from modules.prices.services import PricesService
+from modules.sse.services import SSEService
 
 
 class PricesContainer(containers.DeclarativeContainer):
     yfinance_repository = providers.Singleton(YfinanceRepository)
     prices_service = providers.Singleton(PricesService, repository=yfinance_repository)
-    live_prices = providers.Singleton(LivePricesService)
+    sse_service = providers.Singleton(SSEService)
